requirements.txt
# Base Framework
Django>=4.2,<5.2

# Database Connection (Supabase/Postgres)
psycopg2-binary
dj-database-url
python-dotenv
supabase
djangorestframework

# Data Processing & Excel
pandas
openpyxl
XlsxWriter
django-import-export

# Production Server
gunicorn

# Development Tools
django-debug-toolbar

# UI Framework
django-bootstrap5

# IA y RAG
pgvector
langchain
langchain-community
sentence-transformers
PyMuPDF
O365
msal
django-allauth